refactor(modules): Extract setParameters
