+++ src/sweetalert2.js
@@ -1,7 +1,8 @@
 import defaultParams, { deprecatedParams } from './utils/params.js'
-import { swalClasses, iconTypes } from './utils/classes.js'
+import { swalClasses } from './utils/classes.js'
 import { formatInputOptions, warn, error, warnOnce, callIfFunction } from './utils/utils.js'
 import * as dom from './utils/dom.js'
+import setParameters from './utils/setParameters.js'
 
 let popupParams = Object.assign({}, defaultParams)
 let queue = []
@@ -22,294 +23,6 @@ const showWarningsForParams = (params) => {
     if (sweetAlert.isDeprecatedParameter(param)) {
       warnOnce(`The parameter "${param}" is deprecated and will be removed in the next major release.`)
     }
-  }
-}
-
-/**
- * Set type, text and actions on popup
- *
- * @param params
- * @returns {boolean}
- */
-const setParameters = (params) => {
-  // Determine if the custom target element is valid
-  if (
-    !params.target ||
-    (typeof params.target === 'string' && !document.querySelector(params.target)) ||
-    (typeof params.target !== 'string' && !params.target.appendChild)
-  ) {
-    warn('Target parameter is not valid, defaulting to "body"')
-    params.target = 'body'
-  }
-
-  let popup
-  const oldPopup = dom.getPopup()
-  let targetElement = typeof params.target === 'string' ? document.querySelector(params.target) : params.target
-  // If the model target has changed, refresh the popup
-  if (oldPopup && targetElement && oldPopup.parentNode !== targetElement.parentNode) {
-    popup = dom.init(params)
-  } else {
-    popup = oldPopup || dom.init(params)
-  }
-
-  // Set popup width
-  if (params.width) {
-    popup.style.width = (typeof params.width === 'number') ? params.width + 'px' : params.width
-  }
-
-  // Set popup padding
-  if (params.padding) {
-    popup.style.padding = (typeof params.padding === 'number') ? params.padding + 'px' : params.padding
-  }
-
-  // Set popup background
-  if (params.background) {
-    popup.style.background = params.background
-  }
-  const popupBackgroundColor = window.getComputedStyle(popup).getPropertyValue('background-color')
-  const successIconParts = popup.querySelectorAll('[class^=swal2-success-circular-line], .swal2-success-fix')
-  for (let i = 0; i < successIconParts.length; i++) {
-    successIconParts[i].style.backgroundColor = popupBackgroundColor
-  }
-
-  const container = dom.getContainer()
-  const title = dom.getTitle()
-  const content = dom.getContent().querySelector('#' + swalClasses.content)
-  const actions = dom.getActions()
-  const confirmButton = dom.getConfirmButton()
-  const cancelButton = dom.getCancelButton()
-  const closeButton = dom.getCloseButton()
-  const footer = dom.getFooter()
-
-  // Title
-  if (params.titleText) {
-    title.innerText = params.titleText
-  } else if (params.title) {
-    title.innerHTML = params.title.split('\n').join('<br />')
-  }
-
-  if (typeof params.backdrop === 'string') {
-    dom.getContainer().style.background = params.backdrop
-  } else if (!params.backdrop) {
-    dom.addClass([document.documentElement, document.body], swalClasses['no-backdrop'])
-  }
-
-  // Content as HTML
-  if (params.html) {
-    dom.parseHtmlToContainer(params.html, content)
-
-  // Content as plain text
-  } else if (params.text) {
-    content.textContent = params.text
-    dom.show(content)
-  } else {
-    dom.hide(content)
-  }
-
-  // Position
-  if (params.position in swalClasses) {
-    dom.addClass(container, swalClasses[params.position])
-  } else {
-    warn('The "position" parameter is not valid, defaulting to "center"')
-    dom.addClass(container, swalClasses.center)
-  }
-
-  // Grow
-  if (params.grow && typeof params.grow === 'string') {
-    let growClass = 'grow-' + params.grow
-    if (growClass in swalClasses) {
-      dom.addClass(container, swalClasses[growClass])
-    }
-  }
-
-  // Animation
-  if (typeof params.animation === 'function') {
-    params.animation = params.animation.call()
-  }
-
-  // Close button
-  if (params.showCloseButton) {
-    closeButton.setAttribute('aria-label', params.closeButtonAriaLabel)
-    dom.show(closeButton)
-  } else {
-    dom.hide(closeButton)
-  }
-
-  // Default Class
-  popup.className = swalClasses.popup
-  if (params.toast) {
-    dom.addClass([document.documentElement, document.body], swalClasses['toast-shown'])
-    dom.addClass(popup, swalClasses.toast)
-  } else {
-    dom.addClass(popup, swalClasses.modal)
-  }
-
-  // Custom Class
-  if (params.customClass) {
-    dom.addClass(popup, params.customClass)
-  }
-
-  // Progress steps
-  let progressStepsContainer = dom.getProgressSteps()
-  let currentProgressStep = parseInt(params.currentProgressStep === null ? sweetAlert.getQueueStep() : params.currentProgressStep, 10)
-  if (params.progressSteps && params.progressSteps.length) {
-    dom.show(progressStepsContainer)
-    dom.empty(progressStepsContainer)
-    if (currentProgressStep >= params.progressSteps.length) {
-      warn(
-        'Invalid currentProgressStep parameter, it should be less than progressSteps.length ' +
-        '(currentProgressStep like JS arrays starts from 0)'
-      )
-    }
-    params.progressSteps.forEach((step, index) => {
-      let circle = document.createElement('li')
-      dom.addClass(circle, swalClasses.progresscircle)
-      circle.innerHTML = step
-      if (index === currentProgressStep) {
-        dom.addClass(circle, swalClasses.activeprogressstep)
-      }
-      progressStepsContainer.appendChild(circle)
-      if (index !== params.progressSteps.length - 1) {
-        let line = document.createElement('li')
-        dom.addClass(line, swalClasses.progressline)
-        if (params.progressStepsDistance) {
-          line.style.width = params.progressStepsDistance
-        }
-        progressStepsContainer.appendChild(line)
-      }
-    })
-  } else {
-    dom.hide(progressStepsContainer)
-  }
-
-  // Icon
-  const icons = dom.getIcons()
-  for (let i = 0; i < icons.length; i++) {
-    dom.hide(icons[i])
-  }
-  if (params.type) {
-    let validType = false
-    for (let iconType in iconTypes) {
-      if (params.type === iconType) {
-        validType = true
-        break
-      }
-    }
-    if (!validType) {
-      error(`Unknown alert type: ${params.type}`)
-      return false
-    }
-    const icon = popup.querySelector(`.${swalClasses.icon}.${iconTypes[params.type]}`)
-    dom.show(icon)
-
-    // Animate icon
-    if (params.animation) {
-      dom.addClass(icon, `swal2-animate-${params.type}-icon`)
-    }
-  }
-
-  // Custom image
-  const image = dom.getImage()
-  if (params.imageUrl) {
-    image.setAttribute('src', params.imageUrl)
-    image.setAttribute('alt', params.imageAlt)
-    dom.show(image)
-
-    if (params.imageWidth) {
-      image.setAttribute('width', params.imageWidth)
-    } else {
-      image.removeAttribute('width')
-    }
-
-    if (params.imageHeight) {
-      image.setAttribute('height', params.imageHeight)
-    } else {
-      image.removeAttribute('height')
-    }
-
-    image.className = swalClasses.image
-    if (params.imageClass) {
-      dom.addClass(image, params.imageClass)
-    }
-  } else {
-    dom.hide(image)
-  }
-
-  // Cancel button
-  if (params.showCancelButton) {
-    cancelButton.style.display = 'inline-block'
-  } else {
-    dom.hide(cancelButton)
-  }
-
-  // Confirm button
-  if (params.showConfirmButton) {
-    dom.removeStyleProperty(confirmButton, 'display')
-  } else {
-    dom.hide(confirmButton)
-  }
-
-  // Actions (buttons) wrapper
-  if (!params.showConfirmButton && !params.showCancelButton) {
-    dom.hide(actions)
-  } else {
-    dom.show(actions)
-  }
-
-  // Edit text on confirm and cancel buttons
-  confirmButton.innerHTML = params.confirmButtonText
-  cancelButton.innerHTML = params.cancelButtonText
-
-  // ARIA labels for confirm and cancel buttons
-  confirmButton.setAttribute('aria-label', params.confirmButtonAriaLabel)
-  cancelButton.setAttribute('aria-label', params.cancelButtonAriaLabel)
-
-  // Add buttons custom classes
-  confirmButton.className = swalClasses.confirm
-  dom.addClass(confirmButton, params.confirmButtonClass)
-  cancelButton.className = swalClasses.cancel
-  dom.addClass(cancelButton, params.cancelButtonClass)
-
-  // Buttons styling
-  if (params.buttonsStyling) {
-    dom.addClass([confirmButton, cancelButton], swalClasses.styled)
-
-    // Buttons background colors
-    if (params.confirmButtonColor) {
-      confirmButton.style.backgroundColor = params.confirmButtonColor
-    }
-    if (params.cancelButtonColor) {
-      cancelButton.style.backgroundColor = params.cancelButtonColor
-    }
-
-    // Loading state
-    const confirmButtonBackgroundColor = window.getComputedStyle(confirmButton).getPropertyValue('background-color')
-    confirmButton.style.borderLeftColor = confirmButtonBackgroundColor
-    confirmButton.style.borderRightColor = confirmButtonBackgroundColor
-  } else {
-    dom.removeClass([confirmButton, cancelButton], swalClasses.styled)
-
-    confirmButton.style.backgroundColor = confirmButton.style.borderLeftColor = confirmButton.style.borderRightColor = ''
-    cancelButton.style.backgroundColor = cancelButton.style.borderLeftColor = cancelButton.style.borderRightColor = ''
-  }
-
-  // Footer
-  dom.parseHtmlToContainer(params.footer, footer)
-
-  // CSS animation
-  if (params.animation === true) {
-    dom.removeClass(popup, swalClasses.noanimation)
-  } else {
-    dom.addClass(popup, swalClasses.noanimation)
-  }
-
-  // showLoaderOnConfirm && preConfirm
-  if (params.showLoaderOnConfirm && !params.preConfirm) {
-    warn(
-      'showLoaderOnConfirm is set to true, but preConfirm is not defined.\n' +
-      'showLoaderOnConfirm should be used together with preConfirm, see usage example:\n' +
-      'https://sweetalert2.github.io/#ajax-request'
-    )
   }
 }
